DSH 集成: 4 个项目级 skills + MCP 配置 + clangd 修复 + 调试脚本入库
- .agents/skills/: pecmd-build(构建规程) / msvc-compat(9类症状速查) /
  win-crash-triage(dump解析+探针法, 含 minidump 实测坑) / pecmd-semantics(三重真源法+地址地图)
- .dsh/cordis.yml: memory MCP server (cmd /c npx 中转 Windows .cmd 垫片;
  启动需 dsh --config 显式指定, cordis 不自动发现项目配置)
- compile_flags.txt: 修复 clangd 'pecmd_defs.h' not found 编辑器误报
- tools/debug/: parse_dump.py + thread_rip.py 从临时目录收编入库

+++ src/commands/core_b3l.c
@@ -99,7 +99,7 @@ extern void PECMD_ErrorHandlerWrap(uint8_t *a1, int a2, uint64_t a3, int64_t a4,
 extern void PECMD_ClearFlagAndError(uint8_t *a1);
 extern int PECMD_InstallDevice(LPCWSTR a1, uint32_t a2);
 extern void FUN_14005D534(void *a1, LPCWSTR a2, LPCWSTR a3);
-extern void FUN_14001E6BC(void *a1, LPCWSTR a2, LPCWSTR a3, int a4);
+extern void FUN_14001E6BC(void *a1, LPCWSTR a2, LPCWSTR a3, int64_t a4); /* T1: int→int64_t 与定义一致 */
 extern void *PECMD_LoadIcon(LPCWSTR a1, uint64_t *a2);
 extern int64_t PECMD_GetConfigEntryByName(LPCWSTR a1, uint64_t *a2);
 extern void FUN_1400F429C(WCHAR **pp, WCHAR ch);
